Updated Group.css Somewhat
Updated Group.css to have a few less components, but the current positioning of the button is preventing the background from appearing on the left side of the header. We also still need to come up with a design for the rest of the website, though that can come later.