Updated ProducerUpdated and Server
Updated ProducerUpdated to fill the DJ and Timeslot selects from the database, as well as the Song List from the database. Subsequently updated Server.js to include functions required to get info. Removed other forms of hardcoding and useless variables/comments. Removed checks for if a playlist exists that are no longer needed.